+++ jgclark.Dashboard/src/react/components/MessageOnlyItem.jsx
@@ -1,7 +1,7 @@
 // @flow
 //--------------------------------------------------------------------------
 // Dashboard React component to show a simple message with no other styling. Called by ItemRow.
-// Last updated 2025-03-30 for v2.2.0.a10
+// Last updated 2025-11-23 for v2.3.0.b15, @jgclark
 //--------------------------------------------------------------------------
 
 import React, { type Node } from 'react'
@@ -24,7 +24,7 @@ const MessageOnlyItem = ({ message,
   closingFAIconClassName = '',
   settingsDialogAnchor = '' }: Props): Node => {
   const { setReactSettings } = useAppContext()
-  const contentClassNameToUse = contentClassName || 'messageOnlyItem'
+  const contentClassNameToUse = contentClassName || 'messageItemRow'
 
   // Handle clicking the gear icon to open the settings dialog
   const handleSettingsLinkClick = () => {
@@ -43,16 +43,16 @@ const MessageOnlyItem = ({ message,
   //----------------------------------------------------------------------
 
   return (
-    <div className="sectionItemRow" data-section-type="">
+    <div className="sectionItemRow">
       {rowIconClassName && (
-        <div>
+        <div className="itemIcon">
           <i className={rowIconClassName}></i>
         </div>
       )}
       <div className="sectionItemContent sectionItem">
         <div className={contentClassNameToUse}>
           {message}{' '}
-          {settingsDialogAnchor && <i className="fa-solid fa-gear" onClick={handleSettingsLinkClick}></i>}{' '}
+          {settingsDialogAnchor && <i className="fa-solid fa-gear clickTarget" onClick={handleSettingsLinkClick}></i>}{' '}
           {closingFAIconClassName && <i className={closingFAIconClassName}></i>}
         </div>
       </div>